+++ src/pages/index.tsx
@@ -7,7 +7,6 @@ import { observer } from 'mobx-react-lite';
 import WorkspaceMission from 'people/widgetViews/WorkspaceMission';
 import PeopleHeader from '../people/main/Header';
 import TokenRefresh from '../people/utils/TokenRefresh';
-import AddRepos from '../../src/people/widgetViews/workspace/AddRepositories';
 import BotsBody from './bots/Body';
 import Body from './tribes/Body';
 import Header from './tribes/Header';
@@ -49,10 +48,6 @@ const modeDispatchPages: Record<AppMode, () => React.ReactElement> = {
           <Route path="/workspace/:uuid">
             <WorkspaceMission />
           </Route>
-          <Route path="/workspaces/:uuid">
-            <AddRepos />
-          </Route>
-
           <Route path="/leaderboard">
             <LeaderboardPage />
           </Route>
